+++ lua/obsidian/config/init.lua
@@ -315,6 +315,7 @@ See: https://github.com/obsidian-nvim/obsidian.nvim/wiki/Keymaps]]
   opts.open = tbl_override(defaults.open, opts.open, { schemes = true })
   opts.checkbox = tbl_override(defaults.checkbox, opts.checkbox)
   opts.comment = tbl_override(defaults.comment, opts.comment)
+  opts.outline = tbl_override(defaults.outline, opts.outline)
   opts.frontmatter = tbl_override(defaults.frontmatter, opts.frontmatter)
   opts.search = tbl_override(defaults.search, opts.search)
   opts.note = tbl_override(defaults.note, opts.note)